Fix getFileSize failing (#145)
* Try to fix getFileSize failing

* Fix file sizes over 2 GB

+++ arm9/source/file_browse.h
@@ -26,11 +26,11 @@
 #include <vector>
 
 struct DirEntry {
-	DirEntry(std::string name, size_t size, bool isDirectory, bool isApp, bool selected = false) : name(name), size(size), isDirectory(isDirectory), isApp(isApp), selected(selected) {}
+	DirEntry(std::string name, off_t size, bool isDirectory, bool isApp, bool selected = false) : name(name), size(size), isDirectory(isDirectory), isApp(isApp), selected(selected) {}
 	DirEntry() {}
 
 	std::string name;
-	int size;
+	off_t size;
 	bool isDirectory;
 	bool isApp;
 	bool selected = false;
